test: re-organize test suites to math the same filename and folder
