+++ package.json
@@ -36,13 +36,13 @@
     "transform": {
       ".(ts)": "<rootDir>/node_modules/ts-jest/preprocessor.js"
     },
-    "testRegex": "(Test|spec).ts?$",
+    "testRegex": "(.spec).ts?$",
     "moduleFileExtensions": [
       "ts",
       "js"
     ],
     "setupFiles": [
-      "<rootDir>/spec/helpers/imports.ts"
+      "<rootDir>/spec/setup.ts"
     ],
     "collectCoverageFrom": [
       "src/**/*.ts"
