--- spec/support/jasmine.json
@@ -1,11 +1,0 @@
-{
-  "spec_dir": ".temp/spec",
-  "spec_files": [
-    "**/*[tT]est.js"
-  ],
-  "helpers": [
-    "helpers/**/*.js"
-  ],
-  "stopSpecOnExpectationFailure": false,
-  "random": false
-}
