+++ src/internal/requestAdapter.spec.ts
@@ -1,6 +1,6 @@
 // tslint:disable:max-line-length
-import { MESSAGE } from "../src/config/message";
-import { IHTTPResponse, IRequestAdapter, IRequestOptions, Methods, RequestAdapter } from "../src/internal/requestAdapter";
+import { MESSAGE } from "../config/message";
+import { IHTTPResponse, IRequestAdapter, IRequestOptions, Methods, RequestAdapter } from "./requestAdapter";
 
 /* Mock request adapter */
 export const mockRequestAdapter: IRequestAdapter = {
